Main Backend now handles CORS, Serves Site on basic Route, Tests are now in proper package, Handlers package renamed handler
main.go now starts up the server a little differently now. Using the normal http packages listen serve, CORS handling has been added (As in Dobra's shared setup)

All Routes have been preserved, with the modification of adding "/api" to the front.

The final route "localhost:9000/" now serves the website to the browser.

Package handlers renamed handler to avoid conflict with gorilla/handlers required for CORS fix.

Test folder is now in handler. Probably was working as is before but read up on golang testing and found:
"The test file can be in the same package as the one being tested, or in a corresponding package with the suffix "_test"."

Now reside in handler_test package, which is black-box testing of the functions. Can be changed to handler if anything ever has to be exported into the tests.