Login requests and some CSS Updates
Firstly, added most of Nate's updates to the CSS I missed from the main branch last time. There is one missing because for some reason the new version for main made certain components disappear on my end, so I'll investigate that further later. 2nd big thing is login request on the backend connect service now can take loginInfo as a parameter and send a request, some updates just have to be made to the login form soon to actually use that input.