Settings Progress
I don't want to save to user pages branch yet because there's still a critical issue with the put method and if turns out its more detrimental I don't want to ruin any other code. But I also wanted to make clear the progress I had now. This is the base of the settings component, which right now allows you to change any of your user information and will save the changes to local storage (Where I left off and am going back in the morning is getting to update the database as well. Right now it completely clears the existing values but doesn't add the new ones).