Profile and Setting Pages
Profile pages now exist, accessed by URL using ID number. Divided into three panels of user's info, top goals, and mutual friends. Screen changes if the profile is that of the logged in user. Also now a settings page exists to change info such as username or email. This changes it while logged in but still needs to be properly set up with HTTP requests to put to the database. I'm probably going to do another commit to add cypress tests in a little bit just in case you're seeing this right now.