Cypress tests
Not fully done with adding extra cypress tests, but wanted to this here since I have to come back to it later. This fixes the original cypress tests since none of them worked with the new formatted website, tests for additional navigation through all the new pages, as well as has tests for the profile page component. Also a few pages have slight changes to have IDs or names on elements just so cypress can access them. Still to come is cypress tests on settings component and goals component.