AWS S3 File System for avatar files
Two new routes have been added.

/users/{id}/avatar
put: Updates the user's avatar, takes in a string which needs to be the image's base64 representation

get: returns the base64 representation of the avatar to be displayed on the webpage.

To use this, a .env file with the AWS keys must be added to the httpd directory. Sent in the discord.